Refactor code for improved clarity and functionality
This commit modifies multiple files to improve code clarity and functionality. API class is updated to accept authorization_token as a constructor parameter which promotes instantiation of the API with immediate access to authorized routes. Changes also include renaming indicator constants for more explicit wording and expanding the functionality of valuation and cadastre methods for more flexibility and ease of use. Lastly, unnecessary code and methods are removed for a cleaner implementation.